Add more data
p.s. I had to reclone this repo because when I was committing this last,
I had a power outage for a few seconds and my desktop lost power as it
was committing. In the 4 second window the commit command was running,
my pc lost power. What are the odds.